Update liff.isLoggedIn().html
What We Changed:
liff.isLoggedIn(): After liff.init(), we now use this function to check if the user is already logged in.

liff.login(): If liff.isLoggedIn() is false, we call this function. It will automatically redirect the user to a LINE login and permissions screen. After they log in and grant permission, they will be redirected back to your LIFF app, and the page will reload.

else block: If the user is already logged in, the code proceeds to get their profile, just like before.